backend/requirements.txt
flask==2.2.2
werkzeug==2.2.2
requests
libsql-client
python-dotenv
